pyproject.toml
[project]
name = "router-monitor"
version = "0.1.0"
description = "Code to monitor my router"
readme = "README.md"
authors = [
    { name = "jbeattie5768", email = "jbeattie5768@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rich>=13.7.1",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
